fix(agent): keep the plan-draft reader out of zod's reach, and pin the boundary
The last commit gave the rail the server's own reader so the two could not disagree
about whether a run refused. It did that by importing plan-statement.ts, which
imports the shared statement guard, which imports zod v4 — and zod v4 probes
`new Function("")` at module load to decide whether it may compile validators. A
browser under a strict script-src fires securitypolicyviolation even though zod
catches the throw, so every page rendering an agent run took a CSP violation. The
E2E security-headers spec caught it in chromium and webkit; nothing else did.

The commit message for that change said statement-guard is pure "so that import
costs the browser nothing but the module". Its direct imports were checked and the
transitive closure is what the policy measures.

So the reading half moves to plan-draft.ts, which may import nothing but types and
the fence-tag predicate. plan-statement.ts keeps the validation half and the guard
with it. The rail, the verdict and the recorder all read through plan-draft, so the
one-reader property the last commit bought is unchanged.

Measured rather than argued: built both ways and grepped the client chunks for
zod's own `jitless` marker — one chunk with the old import, none with the new. The
e2e/security-headers.spec.ts run that failed in CI passes locally against this
build.

plan-draft-boundary.test.ts pins the allowed imports as an exact set rather than a
prefix rule. "Anything under @/lib/sql" would admit the grammar and span readers,
which are pure today and one commit from not being, and admitting a module for
where it sits is the reasoning that admitted the guard.
